requirements-exact.txt
Flask==3.0.3
transformers==4.41.2
torch==2.2.2
numpy<2
flask-cors>=3.0
openai>=1.0.0
qdrant-client==1.15.1 
python-dotenv>=1.0.0